리뷰, qna, s3연동

+++ back/src/main/java/com/admin/back/dto/AnswerDto.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class AnswerDto {
-    private int answerId;
+    private Long answerId;
     private String responseText;
     private Date responseDate;
 }
