+++ back/src/main/java/com/admin/back/dto/ReviewResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 @Getter @Setter
 public class ReviewResponseDto {
-    private int responseId;
+    private Long responseId;
     private int adminId;
     private String responseText;
     private Date responseDate;
